All Packets Stored in Database
Packet version 5, 6, 7, 8, and 9 now store information into the
database. Packets 8 and 9 are stored as BLOBs.
id keys are now unsigned ints, and the corresponding fields in the
application are now longs.

+++ NetFlow/src/netflow/PacketV1.java
@@ -38,7 +38,7 @@ public class PacketV1 extends Thread
     }
     
     /*
-     * Grabs and saves all of the packet information.
+     * Grabs all of the packet information and stores it in the database.
      */
     private void SavePacket(DatagramPacket receivedPacket)
     {
@@ -65,7 +65,7 @@ public class PacketV1 extends Thread
                 
                 if (rs != null && rs.next())
                 {
-                    int header_id = rs.getInt(1);
+                    long header_id = rs.getLong(1);
                     
                     for (int i = 0; i < count; i++)
                     {
